+++ recipes/default.rb
@@ -21,7 +21,7 @@ unless node['diptables']['dry_run']
     # set iptables to autolad
     case node['platform_family']
     when 'debian'
-      # TODO: generalize this for other platforms somehow
+        # TODO: generalize this for other platforms somehow
         file '/etc/network/if-up.d/iptables-rules' do
             owner 'root'
             group 'root'
@@ -32,4 +32,9 @@ unless node['diptables']['dry_run']
     else
         Chef::Log.warn("Don't know how to set up automatic iptables on your distribution, sorry. Please submit a bug ticket at https://github.com/wk8/cookbook-iptables/issues")
     end
+end
+
+# flag the node
+ruby_block 'diptables_has_run' do
+    block { node.diptables_has_run true }
 end